requested better error CSVs

+++ src/main.py
@@ -4,9 +4,9 @@ import configparser
 import csv
 import logging
 import sys
-from os import getcwd
+from os import getcwd, remove
 from pathlib import Path
-from typing import List, Set
+from typing import Set
 import shutil
 
 
@@ -27,13 +27,19 @@ class Process:
     # The paths from which to copy
     copy_from_path1: Path = Path()
     copy_from_path2: Path = Path()
+
+    current_id: str = ''
+
+    csv_second_encountered_subids: Set[str] = set()
+
     # We add errored directory names to these CSVs
-    csv_errorfile_main: Path = Path()
-    csv_errorfile_nested: Path = Path()
+    csv_errorfile_first: Path = Path()
+    csv_errorfile_second: Path = Path()
     csv_errorfile_first_writer: csv.writer = None
     csv_errorfile_second_writer: csv.writer = None
     project_homepage: Path = Path()
     individual_gate: Path = Path()
+    logging_level: int = None
     # config parser
     config = configparser.ConfigParser()
 
@@ -71,11 +77,13 @@ class Process:
             self.config.read(self.config_file_path)
 
         # Names are very confusing, but that's how it is sometimes with commissions! :)
+        # Originally these configs were entirely controlled by config file, but then client asked to hardcode defaults
+        # Should probably refactor to saner system but since it's a one-time use thing...
         self.csv_pathfile_first = self.loadPathFromConfig('csv_pathsfiles', 'path_main', default_value='First.csv')
-        self.csv_errorfile_main = self.loadPathFromConfig('csv_errorfiles', 'path_main', nonexistent=True,
-                                                          default_value='First_error.csv')
+        self.csv_errorfile_first = self.loadPathFromConfig('csv_errorfiles', 'path_main', nonexistent=True,
+                                                           default_value='First_error.csv')
         self.csv_pathfile_second = self.loadPathFromConfig('csv_pathsfiles', 'path_nested', default_value='Second.csv')
-        self.csv_errorfile_nested = self.loadPathFromConfig('csv_errorfiles', 'path_nested', nonexistent=True,
+        self.csv_errorfile_second = self.loadPathFromConfig('csv_errorfiles', 'path_nested', nonexistent=True,
                                                             default_value='Second_error.csv')
         self.project_homepage = Path(self.loadValueFromConfig('subdir', 'project_homepage',
                                                               'Project Homepage Attachments'))
@@ -84,12 +92,13 @@ class Process:
         self.work_path = self.loadPathFromConfig('work', 'path', default_value='/temp/project/results/')
         self.copy_from_path1 = self.loadPathFromConfig('copyfrom', 'path1', default_value='/temp/project/temp1/')
         self.copy_from_path2 = self.loadPathFromConfig('copyfrom', 'path2', default_value='/temp/project/temp2/')
+        self.logging_level = int(self.loadValueFromConfig('logging', 'log_level', default_value=str(logging.WARNING)))
 
     def setupLogging(self):
         """ Set up default logging object. """
         # noinspection PyArgumentList
         logging.basicConfig(
-            level=logging.INFO,
+            level=self.logging_level,
             format="%(asctime)s [%(levelname)s] %(message)s",
             handlers=[
                 logging.FileHandler(self.log_file_name),
@@ -99,15 +108,19 @@ class Process:
 
     def openErrorCSVs(self):
         """ Open CSVs and create writer objects. """
-        self.mkdir(self.csv_errorfile_main.parent, parents=True)
-        self.csv_errorfile_first_writer = csv.writer(open(self.csv_errorfile_main, 'w'))
-        self.mkdir(self.csv_errorfile_nested.parent, parents=True)
-        self.csv_errorfile_second_writer = csv.writer(open(self.csv_errorfile_nested, 'w'))
+        if self.csv_errorfile_first.exists():
+            remove(self.csv_errorfile_first)
+        self.mkdir(self.csv_errorfile_first.parent, parents=True)
+        self.csv_errorfile_first_writer = csv.writer(open(self.csv_errorfile_first, 'w'))
+        if self.csv_errorfile_second.exists():
+            remove(self.csv_errorfile_second)
+        self.mkdir(self.csv_errorfile_second.parent, parents=True)
+        self.csv_errorfile_second_writer = csv.writer(open(self.csv_errorfile_second, 'w'))
 
-    @staticmethod
-    def writeRowToErrorCSV(row: List[str], writer_handle: csv.writer):
+    def handleErroredID(self, writer_handle: csv.writer, doc_name, error_type):
         """ Write a column to CSV """
-        writer_handle.writerow(row)
+        writer_handle.writerow([self.current_id, error_type, doc_name])
+        logging.warning(f'Encountered error in "{doc_name}" of type "{error_type}". ID is {self.current_id}')
 
     @staticmethod
     def handleInputCSV(input_csv, row_callback):
@@ -130,28 +143,33 @@ class Process:
             return False
         return True
 
-    @staticmethod
-    def copytree(src: Path, dst: Path):
-        try:
-            logging.info(f'Copying: {src} -> {dst}')
-            shutil.copytree(src, dst, dirs_exist_ok=True)
-        except OSError as os_error:
-            # probably means directory didn't exist
-            logging.warning(f'OSError raised: {os_error} while copying:')
-            logging.warning(f'{src} -> {dst}')
+    def copytree(self, src: Path, dst: Path, writer_handle: csv.writer, doc_name: str):
+        if src.exists():
+            try:
+                logging.info(f'Copying: {src} -> {dst}')
+                shutil.copytree(src, dst, dirs_exist_ok=True)
+            except OSError as os_error:
+                # probably means directory didn't exist
+                logging.info(f'OSError raised: {os_error} while copying:')
+                logging.info(f'{src} -> {dst}')
+                self.handleErroredID(writer_handle=writer_handle,
+                                     doc_name=doc_name, error_type='OS_ERROR')
+        else:
+            self.handleErroredID(writer_handle=writer_handle,
+                                 doc_name=doc_name, error_type='FILE_NOT_EXIST')
 
     def handleRowFirstCSV(self, row: list):
         """ Handle row in main input CSV. """
-        id_string = row[0]
+        self.current_id = row[0]
         # "First.csv will give the name of one folder to be created under
         # say c:\temp\project For example - c:\temp\project\12354
         # Under this folder you will create 2 more empty new folders.
         # Folder names of these 2 folders is provided by config file.
         # Let the folder names be Project Homepage Attachments and Individual Gate Quest Attachments.
         # These are 2 common folders that will exist in every folder provided by First.csv" -- client
-        if id_string not in self.csv_first_encountered_ids:
-            self.csv_first_encountered_ids.add(id_string)
-            id_save_path = Path(self.work_path / id_string)
+        if self.current_id not in self.csv_first_encountered_ids:
+            self.csv_first_encountered_ids.add(self.current_id)
+            id_save_path = Path(self.work_path / self.current_id)
             for subdir in [self.project_homepage, self.individual_gate]:
                 to_create = id_save_path / subdir
                 logging.info(f'Creating path: {to_create}')
@@ -161,35 +179,48 @@ class Process:
             # When you find the folder in this new path, copy that folder to "Project Homepage Attachments"
             # For example, after copy, it will look like c:\temp\project\12354\Project Homepage Attachments\12354"
             # -- client
-            src: Path = self.copy_from_path1 / id_string
-            dst: Path = id_save_path / self.project_homepage / id_string
-            self.copytree(src, dst)
+            src: Path = self.copy_from_path1 / self.current_id
+            dst: Path = id_save_path / self.project_homepage / self.current_id
+            self.copytree(src, dst, writer_handle=self.csv_errorfile_first_writer, doc_name='FirstCSV')
         else:
-            self.writeRowToErrorCSV([id_string], self.csv_errorfile_first_writer)
+            self.handleErroredID(writer_handle=self.csv_errorfile_first_writer,
+                                 doc_name='FirstCSV', error_type='DUPLICATE_ENTRY')
 
     def handleRowSecondCSV(self, row: list):
         """ Handle row in nested input CSV. """
-        id_string = row[0]
-        if id_string in self.csv_first_encountered_ids and id_string not in self.csv_second_encountered_ids:
-            id_save_path = Path(self.work_path / id_string)
-            # "Now, take the same folder name (12354 from First.csv) and search Second.csv.
-            # For every find, you will find multiple different unique folder names listed in column 2 of Second.csv.
-            # Read those different unique folder names from column2 corresponding to column1
-            # (which is essentially First.csv, also look like numbers)" -- client
+        self.current_id = row[0]
+        if self.current_id in self.csv_first_encountered_ids:
+            if self.current_id not in self.csv_second_encountered_ids:
+                id_save_path = Path(self.work_path / self.current_id)
+                # "Now, take the same folder name (12354 from First.csv) and search Second.csv.
+                # For every find, you will find multiple different unique folder names listed in column 2 of Second.csv.
+                # Read those different unique folder names from column2 corresponding to column1
+                # (which is essentially First.csv, also look like numbers)" -- client
 
-            # "Now, and search these new folders -
-            # each and every one of them in c:\temp\project\temp2 (different path and folders)
-            # When you find these folders,
-            # copy those folders directly into Individual Gate Quest Attachments." -- client
+                # "Now, and search these new folders -
+                # each and every one of them in c:\temp\project\temp2 (different path and folders)
+                # When you find these folders,
+                # copy those folders directly into Individual Gate Quest Attachments." -- client
 
-            if len(row) > 1:
-                for column in row[1:]:
-                    src: Path = self.copy_from_path2 / column
-                    dst: Path = id_save_path / self.individual_gate / column
-                    self.copytree(src, dst)
+                if len(row) > 1:
+                    for column in row[1:]:
+                        self.current_id = column
+                        if self.current_id not in self.csv_second_encountered_subids:
+                            self.csv_second_encountered_subids.add(column)
+                            id_save_path2: Path = Path(column)
+                            src: Path = self.copy_from_path2 / id_save_path2
+                            dst: Path = id_save_path / self.individual_gate / id_save_path2
+                            self.copytree(src, dst, writer_handle=self.csv_errorfile_second_writer,
+                                          doc_name='SecondCSV')
+                        else:
+                            self.handleErroredID(writer_handle=self.csv_errorfile_second_writer,
+                                                 doc_name='Second CSV', error_type='DUPLICATE_SUBID')
+                else:
+                    self.handleErroredID(writer_handle=self.csv_errorfile_second_writer,
+                                         doc_name='SecondCSV', error_type='NO_ENTRIES')
             else:
-                logging.warning(
-                    f'Entry in {self.csv_pathfile_second} with ID column {id_string} does not have subdir columns!')
+                self.handleErroredID(writer_handle=self.csv_errorfile_second_writer,
+                                     doc_name='SecondCSV', error_type='ENTRY_MISSING_FROM_FIRST_CSV')
 
     def main(self):
         """ Run program. """
